requirements.txt
asgiref==3.5.1
certifi==2021.10.8
Django==3.2.13
django-cors-headers==3.12.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
wincertstore==0.2
